Struct rc_borrow::RcBorrow

#[repr(transparent)]pub struct RcBorrow<'a, T: ?Sized> { /* fields omitted */ }

Borrowed version of [Rc].

This type is guaranteed to have the same repr as &T.

Implementations

impl<'a, T: ?Sized> RcBorrow<'a, T>

pub fn upgrade(this: Self) -> Rc<T>

Convert this borrowed pointer into an owned pointer.

pub fn to_weak(this: Self) -> Weak<T>

Convert this borrowed pointer into a weak pointer.

pub fn downgrade(this: Self) -> &'a T

Convert this borrowed pointer into a standard reference.

This gives you a long-lived reference, whereas dereferencing gives a temporary borrow.

pub fn strong_count(this: Self) -> usize

Get the number of strong owning pointers to this allocation.

pub fn weak_count(this: Self) -> usize

Get the number of weak owning pointers to this allocation.

pub fn into_raw(this: Self) -> *const T

Get a raw pointer that can be used with from_raw.

pub unsafe fn from_raw(ptr: *const T) -> Self

Construct a new RcBorrow from a raw pointer.

The raw pointer must have been previously returned by a call to RcBorrow<U>::into_raw or Rc<U>::as_raw where U must have the same size and alignment as T. This is trivially true if U is T. Note that if U is not T, this is a pointer cast (transmute) between the two types, and the types must be transmute-compatible.